pyproject.toml
[tool.poetry]
name = "sky-pro-h-w-8"
version = "0.1.0"
description = ""
authors = ["Alina Davydenko <Alinadavydenko620@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
django = "^5.2.3"
djangorestframework = "^3.16.0"
dotenv = "^0.9.9"
psycopg2-binary = "^2.9.10"
pillow = "^11.2.1"
django-filter = "^25.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
